dcs: Rearrange test files
